don't delete launch flag until after the config message comes

+++ projects/openpos-client-core-lib/src/lib/core/services/session.service.ts
@@ -278,7 +278,7 @@ export class SessionService implements IMessageHandler<any> {
             this.personalization.setTheme(message.theme, false);
         }
 
-        if (!this.deletedLaunchFlg) {
+        if (!this.deletedLaunchFlg && message && message.type === 'ConfigChanged') {
             this.deleteLaunchingFlg();
             this.deletedLaunchFlg = true;
         }
